Move the notify interface out to service-api
