+++ acs-configdb/package.json
@@ -19,17 +19,14 @@
   },
   "type": "module",
   "dependencies": {
+    "@amrc-factoryplus/rx-util": "^0.0.3",
     "@amrc-factoryplus/service-api": "^0.0.2",
     "@amrc-factoryplus/utilities": "2.0.0",
     "ajv": "^8.10.0",
     "ajv-formats": "^2.1.1",
-    "deep-equal": "^2.2.3",
     "express-openapi-validator": "^4.13.2",
     "json-merge-patch": "^1.0.2",
-    "path-to-regexp": "^6.3.0",
-    "rxjs": "^7.8.1",
-    "uuid": "^10.0.0",
-    "ws": "^8.18.0"
+    "rxjs": "^7.8.1"
   },
   "devDependencies": {
     "@eslint/js": "^9.13.0",
